Moved progress_indicator to utility
